Add new numeration and use of relation file

+++ utils.py
@@ -5,6 +5,7 @@ import random
 
 from tqdm import tqdm_notebook
 
+import numpy as np
 import pandas as pd
 
 import pydicom
@@ -47,35 +48,87 @@ def move_file(src_filepath, filename, dst_folder, force_extension=None, copy=Tru
         shutil.move(src_filepath, dst_filepath)
 
 
-def organize_folders(src_folder, dst_folder, groups=None, subgroup_length=None, force_extension=None, copy=True, debug=False):
+def organize_folders(src_folder, dst_folder, relation_filepath, reset=False, groups=None, subgroup_length=None, new_numeration=True, filename_prefix='IMG_', force_extension=None, copy=True, debug=False):
     """ Organize folders and files to set all the desired DICOM files into the correct folder """
 
-    # Clean destination folder
-    if os.path.isdir(dst_folder):
-        shutil.rmtree(dst_folder)
+    # In case not reseting the folders, then the current relation is required
+    if not reset:
+        relation_df = open_name_relation_file(relation_filepath, sep=',')
+        if len(relation_df.index) == 0:
+            raise ValueError('No reset is set, but there is no relation file or it is empty')
 
     # Look at all the DICOM files in the source folder, check them and move them appropiatly
     folders = glob(os.path.join(src_folder, '*'))
     correct_filepaths = []
     correct_folders = []
     for folder in tqdm_notebook(folders, desc='Check folders: '):
+        # Find all files in the folder
         filepaths = glob(os.path.join(folder, '*'))
+
+        # Open all the files as DICOM and check if they fullfil the condition to be used in the study
         for filepath in filepaths:
+
+            # If is no resetting and the file is already on the relation, then there is no need to check
+            if not reset:
+                src_path = os.path.split(filepath)[0]
+                if src_path in relation_df.index:
+                    continue
+            
+            # Read and check DICOM
             dcm = pydicom.dcmread(filepath)
             if check_DICOM(dcm, debug):
                 correct_filepaths.append(filepath)
                 correct_folders.append(folder)
 
-    folders_dst_folders = get_final_dst_folder(dst_folder, correct_folders, groups, subgroup_length)
-    for filepath in tqdm_notebook(correct_filepaths, desc='Move files'):
-        path, _ = os.path.split(filepath)
-        final_dst_folder = folders_dst_folders[path]
+    # Only proceed if there is files to move or resetting folders
+    if len(correct_folders) > 0 or reset:
 
-        # Rename the file with the name of the folder (patient ID)
-        _, filename = os.path.split(path)
+        # Relates source folders with destination folder paths depending on shuffle groups and subgroup length
+        folders_dst_folders = get_final_dst_folder(dst_folder, correct_folders, groups, subgroup_length)
+        temp_relation_df = pd.DataFrame(folders_dst_folders.values(), index=folders_dst_folders.keys(), columns=['Path'])
+        temp_relation_df['Filename'] = filename_prefix + str(-1)
+        temp_relation_df.index.rename('Original', inplace=True)
 
-        move_file(filepath, filename, final_dst_folder, force_extension=force_extension, copy=copy)
+        # Define relation DataFrame depending on reset
+        if reset:
+            # Clean destination folder
+            if os.path.isdir(dst_folder):
+                shutil.rmtree(dst_folder)
 
+            relation_df = temp_relation_df
+        else:
+            relation_df = pd.concat([relation_df, temp_relation_df], axis=0)
+
+        # Get last ID of the current files in case of using numeration
+        if new_numeration:
+            current_id = get_last_id(relation_df, prefix=filename_prefix)
+
+        # Loop over the files that should be copied/moved
+        for filepath in tqdm_notebook(correct_filepaths, desc='Move files'):
+
+            # Get the final destination folder
+            src_path, src_filename = os.path.split(filepath)
+            _, src_folder = os.path.split(src_path)
+            final_dst_folder = relation_df.loc[src_path, 'Path']
+
+            # Set filename depending on numeration or patient ID
+            if new_numeration:
+                filename = filename_prefix + str(current_id + 1)
+                current_id += 1
+            else:
+                # Rename the file with the name of the folder (patient ID)
+                filename = src_folder
+
+            # Add new relation on the DataFrame
+            relation_df = add_new_relation(relation_df, src_path, src_filename, filename)
+
+            # Copy/Move the file to the final destination with
+            move_file(filepath, filename, final_dst_folder, force_extension=force_extension, copy=copy)
+
+            # Save the relation file
+            save_name_relation_file(relation_df, relation_filepath, sep=',')
+
+    return relation_df
 
 def expand_list(l, n):
     """ Expand a list `l` to repeat its elements till reaching length of `n` """
@@ -280,3 +333,47 @@ def rename_patient(dicom_files):
         dcm.PatientID = filename
         with open(filepath, 'wb') as f:
             dcm.save_as(f)
+
+
+def open_name_relation_file(filepath, sep=','):
+    """ Extract DataFrame from file containing the relation between original and new files """
+
+    # Check if file exists
+    if not os.path.exists(filepath):
+        df = pd.DataFrame(columns = ['Original', 'New_Path', 'Filename', 'Original_Filename'])
+    else:
+        df = pd.read_csv(filepath, sep=sep, index_col=0)
+
+    return df
+
+
+def save_name_relation_file(relation_df, filepath, sep=','):
+    """ Save file containing the relation between original and new files """
+    relation_df.to_csv(filepath, sep=sep, index=True)
+
+
+def get_last_id(relation_df, prefix='IMG_'):
+    """ Get the ID of the last filename from current relation of files """
+
+    # Extract the maximum ID currently set
+    new_id = relation_df['Filename'].str.split(prefix).str[1].astype(int).max()
+
+    # In case of nan then set it to -1
+    if new_id is np.nan:
+        new_id = -1
+
+    return new_id
+
+
+def add_new_relation(relation_df, src_path, src_filename, new_filename):
+    
+    # Check it does not exist a conflictive addition
+    # if src_path in relation_df.index and 'Original_Filename' in relation_df.columns and not np.isnan(relation_df.loc[src_path, 'Original_Filename']):
+    if src_path in relation_df.index and 'Original_Filename' in relation_df.columns and relation_df['Original_Filename'].notnull()[src_path]:
+        if relation_df.loc[src_path, 'Filename'] != new_filename:
+            raise ValueError(f'For file "{src_path}"" there is already a relation with "{relation_df.loc[src_path, "Filename"]}" but it is being added for "{new_filename}"')
+    else:
+        relation_df.loc[src_path, 'Filename'] = new_filename
+        relation_df.loc[src_path, 'Original_Filename'] = src_filename
+
+    return relation_df
